move barchart folder
move barchart folder to common widgets
