+++ frontend/lib/widgets/common/bar_chart.dart
